modules/anthropic-bedrock-python-ecs-mcp/pyproject.toml
[project]
name = "mcp-sse"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.49.0",
    "aws-cdk-lib>=2.190.0",
    "boto3>=1.38.0",
    "botocore>=1.38.0",
    "constructs>=10.4.2",
    "fastapi>=0.115.12",
    "mcp[cli]>=1.6.0",
]
